Update all client to DependenciesMacros
