+++ App/Anniversaries/Sources/UserDefaultsClient/Client.swift
@@ -5,13 +5,7 @@
 import CoreKit
 import Foundation
 import Dependencies
-
-public struct UserDefaultsClient {
-    public var currentAnniversariesSort: () -> Sort.Kind
-    public var setCurrentAnniversariesSort: (Sort.Kind) -> Void
-    public var currentAnniversariesSortOrder: () -> Sort.Order
-    public var setCurrentAnniversariesSortOrder: (Sort.Order) -> Void
-}
+import DependenciesMacros
 
 extension DependencyValues {
     public var userDefaultsClient: UserDefaultsClient {
@@ -20,19 +14,10 @@ extension DependencyValues {
     }
 }
 
-// MARK: Dependency (testValue, previewValue)
-extension UserDefaultsClient: TestDependencyKey {
-    public static let testValue = Self(
-        currentAnniversariesSort: unimplemented(),
-        setCurrentAnniversariesSort: unimplemented(),
-        currentAnniversariesSortOrder: unimplemented(),
-        setCurrentAnniversariesSortOrder: unimplemented()
-    )
-
-    public static let previewValue = Self(
-        currentAnniversariesSort: { .date },
-        setCurrentAnniversariesSort: { _ in },
-        currentAnniversariesSortOrder: { .ascending },
-        setCurrentAnniversariesSortOrder: { _ in }
-    )
+@DependencyClient
+public struct UserDefaultsClient {
+    public var currentAnniversariesSort: () -> Sort.Kind = { .name }
+    public var setCurrentAnniversariesSort: (Sort.Kind) -> Void
+    public var currentAnniversariesSortOrder: () -> Sort.Order = { .ascending }
+    public var setCurrentAnniversariesSortOrder: (Sort.Order) -> Void
 }
